app/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-network>=24.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-subscription>=3.0.0
gunicorn>=21.2.0
